+++ src/Casters/Caster.php
@@ -11,7 +11,7 @@ abstract class Caster
 {
 	public static array $targets = [];
 	
-	protected static bool $enabled = true;
+	public static bool $enabled = true;
 	
 	public static function register(Application $app): void
 	{
@@ -19,7 +19,7 @@ abstract class Caster
 		
 		foreach (static::$targets as $target) {
 			AbstractCloner::$defaultCasters[$target] = static function($target, array $properties, Stub $stub, bool $is_nested, int $filter = 0) {
-				return static::$enabled
+				return self::$enabled
 					? app(static::class)->cast($target, new Properties($properties), $stub, $is_nested, $filter)
 					: $properties;
 			};
@@ -28,12 +28,12 @@ abstract class Caster
 	
 	public static function disable(): void
 	{
-		static::$enabled = false;
+		self::$enabled = false;
 	}
 	
 	public static function enable(): void
 	{
-		static::$enabled = true;
+		self::$enabled = true;
 	}
 	
 	abstract public function cast($target, Properties $properties, Stub $stub, bool $is_nested, int $filter = 0): array;